+++ config/deps.ts
@@ -11,6 +11,8 @@ export {
 export { oakCors } from "https://deno.land/x/cors@v1.2.2/mod.ts";
 export { Ratelimit } from "https://esm.sh/@upstash/ratelimit@latest";
 export { Redis } from "https://esm.sh/@upstash/redis";
+export { createCipheriv, createDecipheriv, randomBytes } from "node:crypto";
+export { Buffer } from "node:buffer";
 
 export { SemanticCache } from "https://esm.sh/@upstash/semantic-cache";
 export { Index } from "https://esm.sh/@upstash/vector";